dom/workers/test/serviceworkers/test_fetch_event_client_postmessage.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1158735 - FetchEvent.client asserting in onFetch when there's no document.</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>
  <script class="testbody" type="text/javascript">
    SimpleTest.requestCompleteLog();

    var registration;
    function start() {
      return navigator.serviceWorker.register("fetch_event_client.js", { scope: "./"})
        .then((swr) => registration = swr);
    }

    function unregister() {
      return registration.unregister().then(function(result) {
        ok(result, "Unregister should return true.");
      }, function(e) {
        dump("Unregistering the SW failed with " + e + "\n");
      });
    }

    function testFetchEvent() {
      var p = new Promise(function(resolve, reject) {
        var content = document.getElementById("content");
        ok(content, "parent exists.");

        var iframe = document.createElement('iframe');
        iframe.setAttribute('src', "sw_clients/dummy.html");
        content.appendChild(iframe);

        var w = iframe.contentWindow;
        w.navigator.serviceWorker.onmessage = function(msg) {
          ok(msg.data.status, "Receive message posted by client successfully");

          resolve();
        }
      });

      return p;
    }

    function runTest() {
      start()
        .then(testFetchEvent)
        .then(unregister)
        .then(function() {
        }).catch(function(e) {
          ok(false, "Some test failed with error " + e);
        }).then(SimpleTest.finish);
    }

    SimpleTest.waitForExplicitFinish();
    SpecialPowers.pushPrefEnv({"set": [
      ["dom.serviceWorkers.exemptFromPerDomainMax", true],
      ["dom.serviceWorkers.enabled", true],
      ["dom.serviceWorkers.testing.enabled", true],
      ["dom.serviceWorkers.testing.enabled", true],
      ['dom.serviceWorkers.interception.enabled', true],
    ]}, runTest);
  </script>
</body>
</html>
